+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
-## Unreleased changes (2026-05-10)
+## Unreleased changes (2026-05-11)
 
+* feat: raw-chunk training data + parallel load/compress [[details](https://github.com/cbboyan/solverpy/commit/) | [browse](https://github.com/cbboyan/solverpy/tree/)]
 * feat: chunked NPZ training data for 100 GB+ SVM-Light files [[details](https://github.com/cbboyan/solverpy/commit/6504297) | [browse](https://github.com/cbboyan/solverpy/tree/6504297)]
 * fix: Python 3.14 multiprocessing + ENIGMATIC_ROOT model path issues [[details](https://github.com/cbboyan/solverpy/commit/0424a31) | [browse](https://github.com/cbboyan/solverpy/tree/0424a31)]
 * fix: remove trailing commas from parser.add_argument calls in solverpy-autotune [[details](https://github.com/cbboyan/solverpy/commit/3c696bf) | [browse](https://github.com/cbboyan/solverpy/tree/3c696bf)]